refactor: admin 폴더 내부에서 도메인 별로 분리하도록 변경
