+++ src/main/java/com/ziine/admin/artwork/domain/repository/ArtworkStatusHistoryRepository.java
@@ -1,6 +1,6 @@
-package com.ziine.admin.domain.repository;
+package com.ziine.admin.artwork.domain.repository;
 
-import com.ziine.admin.domain.entity.ArtworkStatusHistoryEntity;
+import com.ziine.admin.artwork.domain.entity.ArtworkStatusHistoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
