+++ src/main/java/com/ziine/admin/artwork/presentation/AdminArtworkRetrieveController.java
@@ -1,7 +1,7 @@
-package com.ziine.admin.presentation;
+package com.ziine.admin.artwork.presentation;
 
-import com.ziine.admin.application.AdminArtworkRetrieveService;
-import com.ziine.admin.application.dto.response.AdminArtworkRetrieveResponseDto;
+import com.ziine.admin.artwork.application.AdminArtworkRetrieveService;
+import com.ziine.admin.artwork.application.dto.response.AdminArtworkRetrieveResponseDto;
 import com.ziine.artwork.domain.entity.ArtworkStatus;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
