+++ src/main/java/com/ziine/admin/auth/domain/Admin.java
@@ -1,4 +1,4 @@
-package com.ziine.admin.domain;
+package com.ziine.admin.auth.domain;
 
 public record Admin(
     String name
